Add more models and metrics

+++ backend/src/backend/metrics/mean_directional_accuracy.py
@@ -6,7 +6,7 @@ def mean_directional_accuracy(
     y_hats: np.ndarray,
     *,
     q50_idx: int,
-) -> float:
+) -> np.float64:
     trues_dir = np.sign(y_trues[1:] - y_trues[:-1])
     hats_dir = np.sign(y_hats[1:, :, q50_idx] - y_trues[:-1])
     return (trues_dir == hats_dir).mean()
